feat: Add menu screen and update navigation in home screen
The code changes include adding a new menu screen and updating the navigation in the home screen to navigate to the menu screen instead of the login screen when tapped. This improves the user experience by providing a direct way to access the menu.

Recent user commits:
- Fix inactivity timer duration in login_screen.dart
- Add fluttertoast dependency, handle error in fetchFaturas, and add INTERNET permission in AndroidManifest.xml

Recent repository commits:
- Fix inactivity timer duration in login_screen.dart
- Add fluttertoast dependency, handle error in fetchFaturas, and add INTERNET permission in AndroidManifest.xml
- Initial Commit